Galactic parallelization and new parts flag
The galactic part now performs a parallel propagation (a for loop is no longer needed). Additionally, we added a new flag (--parts) to divide the simulation into n parts. This way, we avoid memory issues as a consecuence of the use of ParticleCollector (see crpropa documentation for further information). Finally, a new functionality to handle different celestial coordinates was added, more details in future commits.