Removing garbage output
The garbage observer was introduced to speed up the simulations by
stopping particles not in the direction of the main observer. The
problem with this was the creation of very large output files initially
thought for debugging purposes. Now I am removing this output, so the
simulation just deletes particles detected by this "test" observer.
Additionally, I am using f-strings now when necessary.