Smart checkpoint handling
Added the class `checkpoint_state` to correctly handling checkpoints
both for text and pickle files. The text files encodes the current
generation in their name, while the pickle files saves an object of the
`checkpoint_state`, being `igen` one of its attributes. This attribute
correspond to the current generation.
Additionally, a function for sorting checkpoint files by generation was
added. This function uses regex to extract the generation from the filename.
Finally, at the end of each generation, the checkpoint state is dumped
into a pickle file, saving the state after the random mutation, for
continuation purposes (notice that the text files are saved before this
mutation and only if there is a change in the top N most performant
solutions).